feature: improve extension api tree shaking
