+++ packages/extension-api/package.json
@@ -27,6 +27,7 @@
   ],
   "sideEffects": false,
   "dependencies": {
+    "@lvce-editor/command": "^2.0.0",
     "@lvce-editor/rpc": "^6.4.0",
     "@lvce-editor/rpc-registry": "^9.24.0",
     "@lvce-editor/virtual-dom-worker": "^10.0.0"
