+++ packages/extension-api/src/parts/CommandMap/CommandMap.ts
@@ -1,12 +1,16 @@
-import { commandMap as extensionApiCommandMap } from '../ExtensionApiCommandMap/ExtensionApiCommandMap.ts'
+import * as CoreExtensionApiCommandMap from '../CoreExtensionApiCommandMap/CoreExtensionApiCommandMap.ts'
+import * as ExtensionApiCommandRegistry from '../ExtensionApiCommandRegistry/ExtensionApiCommandRegistry.ts'
 import { handleExtensionManagementMessagePort } from '../ExtensionApiWorkerHandleMessagePort/ExtensionApiWorkerHandleMessagePort.ts'
 
-export const commandMap = {
-  ...extensionApiCommandMap,
-  async initialize(type: string, port: MessagePort): Promise<void> {
-    if (type !== 'message-port') {
-      throw new Error(`unsupported initialize type ${type}`)
-    }
-    await handleExtensionManagementMessagePort(port)
-  },
+export const getCommandMap = (): ExtensionApiCommandRegistry.ExtensionApiCommandMap => {
+  return {
+    ...CoreExtensionApiCommandMap.commandMap,
+    ...ExtensionApiCommandRegistry.getCommandMap(),
+    async initialize(type: string, port: MessagePort): Promise<void> {
+      if (type !== 'message-port') {
+        throw new Error(`unsupported initialize type ${type}`)
+      }
+      await handleExtensionManagementMessagePort(port)
+    },
+  }
 }
