+++ packages/extension-api/src/parts/CoreExtensionApiCommandMap/CoreExtensionApiCommandMap.ts
@@ -1,0 +1,7 @@
+const getStatusBarItems = (): readonly never[] => {
+  return []
+}
+
+export const commandMap = {
+  'ExtensionApi.getStatusBarItems': getStatusBarItems,
+}
