+++ packages/extension-api/src/parts/ExtensionApiWorkerCommandMap/ExtensionApiWorkerCommandMap.ts
@@ -1,13 +1,17 @@
-import { commandMap as extensionApiCommandMap } from '../ExtensionApiCommandMap/ExtensionApiCommandMap.ts'
+import * as CoreExtensionApiCommandMap from '../CoreExtensionApiCommandMap/CoreExtensionApiCommandMap.ts'
+import * as ExtensionApiCommandRegistry from '../ExtensionApiCommandRegistry/ExtensionApiCommandRegistry.ts'
 import { handleExtensionManagementMessagePort } from '../ExtensionApiWorkerHandleMessagePort/ExtensionApiWorkerHandleMessagePort.ts'
 
-export const commandMap = {
-  ...extensionApiCommandMap,
-  'HandleMessagePort.handleExtensionManagementMessagePort': handleExtensionManagementMessagePort,
-  async initialize(type: string, port: MessagePort): Promise<void> {
-    if (type !== 'message-port') {
-      throw new Error(`unsupported initialize type ${type}`)
-    }
-    await handleExtensionManagementMessagePort(port)
-  },
+export const getCommandMap = (): ExtensionApiCommandRegistry.ExtensionApiCommandMap => {
+  return {
+    ...CoreExtensionApiCommandMap.commandMap,
+    ...ExtensionApiCommandRegistry.getCommandMap(),
+    'HandleMessagePort.handleExtensionManagementMessagePort': handleExtensionManagementMessagePort,
+    async initialize(type: string, port: MessagePort): Promise<void> {
+      if (type !== 'message-port') {
+        throw new Error(`unsupported initialize type ${type}`)
+      }
+      await handleExtensionManagementMessagePort(port)
+    },
+  }
 }
