--- packages/extension-api/test/parts/ExtensionApiCommandMap/ExtensionApiCommandMap.test.ts
@@ -1,49 +1,0 @@
-import { strictEqual } from 'node:assert/strict'
-import { test } from 'node:test'
-import { commandMap } from '../../../src/parts/ExtensionApiCommandMap/ExtensionApiCommandMap.ts'
-
-test('commandMap exposes hover provider commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.executeHoverProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeLanguageProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeOrganizeImportsProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getHoverProviderRegistrySnapshot'], 'function')
-})
-
-test('commandMap exposes signature help provider commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.executeSignatureHelpProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getSignatureHelpProviderRegistrySnapshot'], 'function')
-})
-
-test('commandMap exposes view commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.executeViewProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getViewActions'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getViewActionsDom'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getViewMenuEntries'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getViewRegistrySnapshot'], 'function')
-})
-
-test('commandMap exposes output channel commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.getOutputChannelRegistrySnapshot'], 'function')
-})
-
-test('commandMap keeps completion, diagnostic and formatting provider commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.executeCompletionProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeDiagnosticProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeFormattingProvider'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeResolveCompletionItemProvider'], 'function')
-})
-
-test('commandMap exposes file system provider commands', () => {
-  strictEqual(typeof commandMap['ExtensionApi.executeFileSystemProviderGetPathSeparator'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeFileSystemProviderIsReadonly'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeFileSystemProviderReadDirWithFileTypes'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.executeFileSystemProviderReadFile'], 'function')
-  strictEqual(typeof commandMap['ExtensionApi.getFileSystemProviderRegistrySnapshot'], 'function')
-})
-
-test('commandMap exposes debug provider commands', () => {
-  strictEqual(typeof commandMap['ExtensionHostDebug.start'], 'function')
-  strictEqual(typeof commandMap['ExtensionHostDebug.pause'], 'function')
-  strictEqual(typeof commandMap['ExtensionHostDebug.resume'], 'function')
-  strictEqual(typeof commandMap['ExtensionHostDebug.evaluate'], 'function')
-})
